+++ kibot/__main__.py
@@ -10,11 +10,12 @@
 Usage:
   kibot [-b BOARD] [-e SCHEMA] [-c CONFIG] [-d OUT_DIR] [-s PRE] [-D]
          [-q | -v...] [-C | -i | -n] [-m MKFILE] [-A] [-g DEF] ...
-         [-E DEF] ... [-w LIST] [TARGET...]
-  kibot [-v...] [-b BOARD] [-e SCHEMA] [-c PLOT_CONFIG] [-E DEF] ... --list
-  kibot [-v...] [-b BOARD] [-d OUT_DIR] [-p | -P] --example
-  kibot [-v...] [--start PATH] [-d OUT_DIR] [--dry] [-t, --type TYPE]...
-         --quick-start
+         [-E DEF] ... [-w LIST] [--banner N] [TARGET...]
+  kibot [-v...] [-b BOARD] [-e SCHEMA] [-c PLOT_CONFIG] [--banner N]
+         [-E DEF] ... --list
+  kibot [-v...] [-b BOARD] [-d OUT_DIR] [-p | -P] [--banner N] --example
+  kibot [-v...] [--start PATH] [-d OUT_DIR] [--dry] [--banner N]
+         [-t, --type TYPE]... --quick-start
   kibot [-v...] --help-filters
   kibot [-v...] [--markdown|--json] --help-dependencies
   kibot [-v...] --help-global-options
@@ -23,6 +24,7 @@ Usage:
   kibot [-v...] --help-outputs
   kibot [-v...] --help-preflights
   kibot [-v...] --help-variants
+  kibot [-v...] --help-banners
   kibot -h | --help
   kibot --version
 
@@ -32,6 +34,7 @@ Arguments:
 Options:
   -A, --no-auto-download           Disable dependencies auto-download
   -b BOARD, --board-file BOARD     The PCB .kicad-pcb board file
+  --banner N                       Display banner number N (-1 == random)
   -c CONFIG, --plot-config CONFIG  The plotting config file to use
   -C, --cli-order                  Generate outputs using the indicated order
   -d OUT_DIR, --out-dir OUT_DIR    The output directory [default: .]
@@ -60,6 +63,7 @@ Quick start options:
 
 Help options:
   -h, --help                       Show this help message and exit
+  --help-banners                   Show all available banners
   --help-dependencies              List dependencies in human readable format
   --help-filters                   List supported filters and details
   --help-global-options            List supported global variables
@@ -97,6 +101,7 @@ if os.environ.get('KIAUS_USE_NIGHTLY'):  # pragma: no cover (nightly)
     else:
         os.environ['PYTHONPATH'] = pcbnew_path
     nightly = True
+from .banner import get_banner, BANNERS
 from .gs import GS
 from . import dep_downloader
 from .misc import EXIT_BAD_ARGS, W_VARCFG, NO_PCBNEW_MODULE, W_NOKIVER, hide_stderr, TRY_INSTALL_CHECK, W_ONWIN
@@ -330,6 +335,14 @@ def main():
     # Load output and preflight plugins
     load_actions()
 
+    if args.banner is not None:
+        try:
+            id = int(args.banner)
+        except ValueError:
+            logger.error('The banner option needs an integer ({})'.format(id))
+            sys.exit(EXIT_BAD_ARGS)
+        logger.info(get_banner(id))
+
     if args.help_outputs or args.help_list_outputs:
         print_outputs_help(details=args.help_outputs)
         sys.exit(0)
@@ -350,6 +363,11 @@ def main():
         sys.exit(0)
     if args.help_dependencies:
         print_dependencies(args.markdown, args.json)
+        sys.exit(0)
+    if args.help_banners:
+        for c, b in enumerate(BANNERS):
+            logger.info('Banner '+str(c))
+            logger.info(b)
         sys.exit(0)
     if args.example:
         check_board_file(args.board_file)